Python script to get information from academic calendar
this script would have to be called and given a different url based off of the time frame being asked for. For example, if I want to know the late drop deadline for fall of 2020, the end of the url would have to be changed to "fall20". From there the script would run and get the information in json format so it can be used internal to Google's systems